update upcoming event dynamic get from database
