+++ .env.example
@@ -1,6 +1,6 @@
-DB_HOST=localhost
+DB_HOST=mysql8
 DB_USER=root
-DB_PASS=
+DB_PASS=pass
 DB_NAME=events
 
 APP_NAME=Sistem Pengelolaan Event HIMATIKA UNSIA
